// frontend/src/app/core/api/model/userDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserDto { 
    id?: number;
    name?: string;
    email?: string;
    role?: UserDtoRoleEnum;
    teamName?: string;
    createdAt?: string;
}
export enum UserDtoRoleEnum {
    Requester = 'REQUESTER',
    ProcurementOfficer = 'PROCUREMENT_OFFICER',
    FinanceOfficer = 'FINANCE_OFFICER',
    Administrator = 'ADMINISTRATOR'
};